requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-bcrypt==1.0.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic>=2.0.0
bcrypt==4.1.1
requests==2.31.0
psycopg2-binary==2.9.9
PyJWT==2.8.0